split evaluator
