+++ evaluator/prefix.go
@@ -1,0 +1,16 @@
+package evaluator
+
+import (
+	"github.com/flipez/rocket-lang/object"
+)
+
+func evalPrefixExpression(operator string, right object.Object) object.Object {
+	switch operator {
+	case "!":
+		return evalBangOperatorExpression(right)
+	case "-":
+		return evalMinusPrefixOperatorExpression(right)
+	default:
+		return newError("unknown operator: %s%s", operator, right.Type())
+	}
+}
